define integers to hold intptr_t rather than long
in anticipation of tagged pointers. also: i hate the built in
c integer types with their vague lengths.

+++ object.h
@@ -74,7 +74,7 @@ typedef struct symbol_s {
 
 typedef struct integer_s {
   struct gc_header header;                  /* TYPE_INTEGER */
-  long integer;                 /* the integer */
+  intptr_t integer;                 /* the integer */
 } integer_s;
 
 typedef struct special_s {
